Add AIR_GAPPED region support (#4)
fixed the additional-packages install to do a yum "localinstall" on all
packages in the directory. Localinstall considers all rpm package files
together as if they are being installed from a repo, and will correctly
order the installation of the rpms if some of them have dependencies on
each other.

+++ scripts/install-ecs-init.sh
@@ -1,6 +1,11 @@
 #!/usr/bin/env bash
 set -ex
 
+if [ -n "$AIR_GAPPED" ]; then
+    echo "Air-gapped region, assuming ecs-init and dependencies will be in additional-packages/ directory"
+    exit 0
+fi
+
 WORK_DIR="$(mktemp -d)"
 trap "rm -rf ${WORK_DIR}" EXIT
